requirements.txt
aiohttp==3.9.1
asyncpg==0.29.0
python-dotenv==1.0.0
certifi==2023.7.22
Flask==3.0.0
psycopg2-binary==2.9.9
requests==2.31.0
